i#7113 instr decode cache: move module mapper into raw2trace_shared
Moves the module mapper and related functionality into raw2trace_shared. This is in preparation
for further changes that include the module mapper logic in a new instr_decode_cache_t. To avoid
having to pull in the whole drmemtrace_raw2trace implementation into the instr decode cache
library, which will end up including it in the invariant checker and everything that depends on,
we separate out the module mapper into raw2trace_shared which is intended for such cases.

Issue: #7113
